Migrating to style-dictionary 4
