+++ libs/tokens/src/actions.js
@@ -1,15 +1,14 @@
 import * as fs from 'fs-extra'
-import { Dictionary, Platform } from 'style-dictionary'
 
 export default {
   copy_package_json: {
     name: 'copy_package_json',
-    do: function (dictionary: Dictionary, config: Platform) {
+    do: function (dictionary, config) {
       console.log('Copying package.json')
       console.log(config)
       fs.copySync('package.json', config.buildPath + '..')
     },
-    undo: function (dictionary: Dictionary, config: Platform) {
+    undo: function (dictionary, config) {
       console.log('Cleaning assets directory')
       console.log(config)
       fs.removeSync(config.buildPath + '../package.json')
