+++ libs/tokens/src/tokens/2023/ref/blue.ref.json
