+++ libs/tokens/src/tokens/2023/ref/neutral.ref.json
