+++ libs/tokens/src/tokens/2023/sys/density.comfortable.json
@@ -1,0 +1,74 @@
+{
+  "sys": {
+    "space": {
+      "4xs": {
+        "value": "1",
+        "type": "dimension"
+      },
+      "3xs": {
+        "value": "2",
+        "type": "dimension"
+      },
+      "2xs": {
+        "value": "4",
+        "type": "dimension"
+      },
+      "xs": {
+        "value": "8",
+        "type": "dimension"
+      },
+      "s": {
+        "value": "12",
+        "type": "dimension"
+      },
+      "m": {
+        "value": "16",
+        "type": "dimension"
+      },
+      "l": {
+        "value": "24",
+        "type": "dimension"
+      },
+      "xl": {
+        "value": "32",
+        "type": "dimension"
+      },
+      "2xl": {
+        "value": "40",
+        "type": "dimension"
+      },
+      "3xl": {
+        "value": "48",
+        "type": "dimension"
+      },
+      "4xl": {
+        "value": "64",
+        "type": "dimension"
+      },
+      "5xl": {
+        "value": "80",
+        "type": "dimension"
+      },
+      "6xl": {
+        "value": "96",
+        "type": "dimension"
+      },
+      "7xl": {
+        "value": "112",
+        "type": "dimension"
+      },
+      "8xl": {
+        "value": "120",
+        "type": "dimension"
+      },
+      "max": {
+        "value": "999",
+        "type": "dimension"
+      },
+      "0": {
+        "value": "0",
+        "type": "dimension"
+      }
+    }
+  }
+}
