+++ libs/tokens/src/tokens/2023/sys/typography/weight.json
