--- libs/tokens/src/tokens/color/sys/accent.json
@@ -1,28 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "accent": {
-        "accentGreen": {
-          "value": "{ref.color.green.40}",
-          "darkValue": "{ref.color.green.40}",
-          "type": "color"
-        },
-        "onAccentGreen": {
-          "value": "{ref.color.white}",
-          "darkValue": "{ref.color.white}",
-          "type": "color"
-        },
-        "accentOrange": {
-          "value": "{ref.color.orange.80}",
-          "darkValue": "{ref.color.orange.80}",
-          "type": "color"
-        },
-        "onAccentOrange": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
