--- libs/tokens/src/tokens/color/sys/background.json
@@ -1,18 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "background": {
-        "background": {
-          "value": "{ref.color.white}",
-          "darkValue": "{ref.color.grey.15}",
-          "type": "color"
-        },
-        "backgroundDim": {
-          "value": "{ref.color.grey.95}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
