--- libs/tokens/src/tokens/color/sys/container.json
@@ -1,63 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "container": {
-        "container": {
-          "value": "{ref.color.grey.95}",
-          "darkValue": "{ref.color.grey.10}",
-          "type": "color"
-        },
-        "containerDim1": {
-          "value": "{ref.color.grey.90}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        },
-        "containerDim2": {
-          "value": "{ref.color.grey.85}",
-          "darkValue": "{ref.color.grey.30}",
-          "type": "color"
-        },
-        "containerBright": {
-          "value": "{ref.color.white}",
-          "darkValue": "{ref.color.grey.10}",
-          "type": "color"
-        },
-        "containerShade1": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.85}",
-          "type": "color"
-        },
-        "containerShade2": {
-          "value": "{ref.color.grey.10}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        },
-        "containerShade3": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        },
-        "containerDisabled": {
-          "value": "{ref.color.grey.98}",
-          "darkValue": "{ref.color.grey.25}",
-          "type": "color"
-        },
-        "containerPositive": {
-          "value": "{ref.color.green.40}",
-          "darkValue": "{ref.color.green.40}",
-          "type": "color"
-        },
-        "containerNegative": {
-          "value": "{ref.color.red.40}",
-          "darkValue": "{ref.color.red.45}",
-          "type": "color"
-        },
-        "containerNegativeBright": {
-          "value": "{ref.color.red.95}",
-          "darkValue": "{ref.color.red.20}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
