--- libs/tokens/src/tokens/color/sys/content.json
@@ -1,53 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "content": {
-        "content": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.90}",
-          "type": "color"
-        },
-        "contentInverse": {
-          "value": "{ref.color.white}",
-          "darkValue": "{ref.color.grey.15}",
-          "type": "color"
-        },
-        "contentSecondary": {
-          "value": "{ref.color.grey.40}",
-          "darkValue": "{ref.color.grey.65}",
-          "type": "color"
-        },
-        "contentPositive": {
-          "value": "{ref.color.green.40}",
-          "darkValue": "{ref.color.green.65}",
-          "type": "color"
-        },
-        "contentPositiveBright": {
-          "value": "{ref.color.green.98}",
-          "darkValue": "{ref.color.green.98}",
-          "type": "color"
-        },
-        "contentNegative": {
-          "value": "{ref.color.red.40}",
-          "darkValue": "{ref.color.red.65}",
-          "type": "color"
-        },
-        "contentNegativeBright": {
-          "value": "{ref.color.red.98}",
-          "darkValue": "{ref.color.red.98}",
-          "type": "color"
-        },
-        "contentCustomBlueBright": {
-          "value": "{ref.color.blue.98}",
-          "darkValue": "{ref.color.blue.98}",
-          "type": "color"
-        },
-        "contentDisabled": {
-          "value": "{ref.color.grey.60}",
-          "darkValue": "{ref.color.grey.80}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
