--- libs/tokens/src/tokens/color/sys/custom.json
@@ -1,68 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "custom": {
-        "customBlue": {
-          "value": "{ref.color.blue.40}",
-          "darkValue": "{ref.color.blue.45}",
-          "type": "color"
-        },
-        "onCustomBlue": {
-          "value": "{ref.color.blue.90}",
-          "darkValue": "{ref.color.blue.85}",
-          "type": "color"
-        },
-        "customBlueBright": {
-          "value": "{ref.color.blue.90}",
-          "darkValue": "{ref.color.blue.60}",
-          "type": "color"
-        },
-        "onCustomBlueBright": {
-          "value": "{ref.color.blue.20}",
-          "darkValue": "{ref.color.blue.90}",
-          "type": "color"
-        },
-        "customGreen": {
-          "value": "{ref.color.green.20}",
-          "darkValue": "{ref.color.green.20}",
-          "type": "color"
-        },
-        "onCustomGreen": {
-          "value": "{ref.color.green.98}",
-          "darkValue": "{ref.color.green.98}",
-          "type": "color"
-        },
-        "customGreenBright": {
-          "value": "{ref.color.green.98}",
-          "darkValue": "{ref.color.green.98}",
-          "type": "color"
-        },
-        "onCustomGreenBright": {
-          "value": "{ref.color.green.20}",
-          "darkValue": "{ref.color.green.20}",
-          "type": "color"
-        },
-        "customGrey": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        },
-        "onCustomGrey": {
-          "value": "{ref.color.grey.90}",
-          "darkValue": "{ref.color.grey.90}",
-          "type": "color"
-        },
-        "customGreyBright": {
-          "value": "{ref.color.grey.90}",
-          "darkValue": "{ref.color.grey.90}",
-          "type": "color"
-        },
-        "onCustomGreyBright": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
