--- libs/tokens/src/tokens/color/sys/primary.json
@@ -1,13 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "primary": {
-        "primary": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
