--- libs/tokens/src/tokens/color/sys/state-layers.json
@@ -1,62 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "state-layers": {
-        "stateBlack": {
-          "value": "{ref.color.black}",
-          "darkValue": "{ref.color.black}",
-          "alpha": 0.1,
-          "type": "color"
-        },
-        "stateBlackDim1": {
-          "value": "{ref.color.black}",
-          "darkValue": "{ref.color.black}",
-          "alpha": 0.2,
-          "type": "color"
-        },
-        "stateBlackDim2": {
-          "value": "{ref.color.black}",
-          "darkValue": "{ref.color.black}",
-          "alpha": 0.4,
-          "type": "color"
-        },
-        "stateBlackShade": {
-          "value": "{ref.color.black}",
-          "darkValue": "{ref.color.black}",
-          "alpha": 0.6,
-          "type": "color"
-        },
-        "statePositive": {
-          "value": "{sys.color.status.positive.positive}",
-          "darkValue": "{sys.color.status.positive.positive}",
-          "alpha": 0.1,
-          "type": "color"
-        },
-        "statePositiveDim": {
-          "value": "{sys.color.status.positive.positive}",
-          "darkValue": "{sys.color.status.positive.positive}",
-          "alpha": 0.2,
-          "type": "color"
-        },
-        "stateNegative": {
-          "value": "{sys.color.status.negative.negative}",
-          "darkValue": "{sys.color.status.negative.negative}",
-          "alpha": 0.1,
-          "type": "color"
-        },
-        "stateNegativeDim": {
-          "value": "{sys.color.status.negative.negative}",
-          "darkValue": "{sys.color.status.negative.negative}",
-          "alpha": 0.2,
-          "type": "color"
-        },
-        "stateCustomBlue": {
-          "value": "{sys.color.custom.customBlue}",
-          "darkValue": "{sys.color.custom.customBlue}",
-          "alpha": 0.2,
-          "type": "color"
-        }
-      }
-    }
-  }
-}
