--- libs/tokens/src/tokens/color/sys/status.json
@@ -1,118 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "status": {
-        "information": {
-          "information": {
-            "value": "{ref.color.grey.20}",
-            "darkValue": "{ref.color.grey.80}",
-            "type": "color"
-          },
-          "onInformation": {
-            "value": "{ref.color.white}",
-            "darkValue": "{ref.color.grey.20}",
-            "type": "color"
-          },
-          "informationBright": {
-            "value": "{ref.color.grey.95}",
-            "darkValue": "{ref.color.grey.20}",
-            "type": "color"
-          },
-          "onInformationBright": {
-            "value": "{ref.color.grey.20}",
-            "darkValue": "{ref.color.grey.95}",
-            "type": "color"
-          }
-        },
-        "negative": {
-          "negative": {
-            "value": "{ref.color.red.40}",
-            "darkValue": "{ref.color.red.70}",
-            "type": "color"
-          },
-          "onNegative": {
-            "value": "{ref.color.white}",
-            "darkValue": "{ref.color.red.95}",
-            "type": "color"
-          },
-          "negativeBright": {
-            "value": "{ref.color.red.95}",
-            "darkValue": "{ref.color.red.20}",
-            "type": "color"
-          },
-          "onNegativeBright": {
-            "value": "{ref.color.red.40}",
-            "darkValue": "{ref.color.red.95}",
-            "type": "color"
-          }
-        },
-        "warning": {
-          "warning": {
-            "value": "{ref.color.orange.50}",
-            "darkValue": "{ref.color.orange.45}",
-            "type": "color"
-          },
-          "onWarning": {
-            "value": "{ref.color.white}",
-            "darkValue": "{ref.color.orange.98}",
-            "type": "color"
-          },
-          "warningBright": {
-            "value": "{ref.color.orange.95}",
-            "darkValue": "{ref.color.orange.20}",
-            "type": "color"
-          },
-          "onWarningBright": {
-            "value": "{ref.color.orange.40}",
-            "darkValue": "{ref.color.orange.95}",
-            "type": "color"
-          }
-        },
-        "positive": {
-          "positive": {
-            "value": "{ref.color.green.40}",
-            "darkValue": "{ref.color.green.45}",
-            "type": "color"
-          },
-          "onPositive": {
-            "value": "{ref.color.white}",
-            "darkValue": "{ref.color.green.98}",
-            "type": "color"
-          },
-          "positiveBright": {
-            "value": "{ref.color.green.98}",
-            "darkValue": "{ref.color.green.20}",
-            "type": "color"
-          },
-          "onPositiveBright": {
-            "value": "{ref.color.green.40}",
-            "darkValue": "{ref.color.green.98}",
-            "type": "color"
-          }
-        },
-        "notice": {
-          "notice": {
-            "value": "{ref.color.blue.40}",
-            "darkValue": "{ref.color.blue.60}",
-            "type": "color"
-          },
-          "onNotice": {
-            "value": "{ref.color.white}",
-            "darkValue": "{ref.color.blue.15}",
-            "type": "color"
-          },
-          "noticeBright": {
-            "value": "{ref.color.blue.95}",
-            "darkValue": "{ref.color.blue.15}",
-            "type": "color"
-          },
-          "onNoticeBright": {
-            "value": "{ref.color.blue.40}",
-            "darkValue": "{ref.color.blue.95}",
-            "type": "color"
-          }
-        }
-      }
-    }
-  }
-}
