--- libs/tokens/src/tokens/color/sys/stroke.json
@@ -1,63 +1,0 @@
-{
-  "sys": {
-    "color": {
-      "stroke": {
-        "stroke": {
-          "value": "{ref.color.grey.20}",
-          "darkValue": "{ref.color.grey.80}",
-          "type": "color"
-        },
-        "strokeVariant1": {
-          "value": "{ref.color.grey.50}",
-          "darkValue": "{ref.color.grey.30}",
-          "type": "color"
-        },
-        "strokeVariant2": {
-          "value": "{ref.color.grey.80}",
-          "darkValue": "{ref.color.grey.20}",
-          "type": "color"
-        },
-        "strokeDisabled": {
-          "value": "{ref.color.grey.60}",
-          "darkValue": "{ref.color.grey.60}",
-          "type": "color"
-        },
-        "strokeNotice": {
-          "value": "{ref.color.blue.40}",
-          "darkValue": "{ref.color.blue.60}",
-          "type": "color"
-        },
-        "strokePositive": {
-          "value": "{ref.color.green.40}",
-          "darkValue": "{ref.color.green.65}",
-          "type": "color"
-        },
-        "strokeWarning": {
-          "value": "{ref.color.orange.40}",
-          "darkValue": "{ref.color.orange.55}",
-          "type": "color"
-        },
-        "strokeNegative": {
-          "value": "{ref.color.red.40}",
-          "darkValue": "{ref.color.red.65}",
-          "type": "color"
-        },
-        "strokeCustomBlue": {
-          "value": "{ref.color.blue.40}",
-          "darkValue": "{ref.color.blue.60}",
-          "type": "color"
-        },
-        "strokeCustomBlueBright": {
-          "value": "{ref.color.blue.70}",
-          "darkValue": "{ref.color.blue.20}",
-          "type": "color"
-        },
-        "strokeInversed": {
-          "value": "{ref.color.white}",
-          "darkValue": "{ref.color.black}",
-          "type": "color"
-        }
-      }
-    }
-  }
-}
