--- libs/tokens/src/tokens/motion/motion.json
@@ -1,46 +1,0 @@
-{
-  "sys": {
-    "motion": {
-      "easing": {
-        "ease-in": {
-          "value": "cubic-bezier(0.64, 0, 0.78, 0)",
-          "type": "string"
-        },
-        "ease-in-out": {
-          "value": "cubic-bezier(0.83, 0, 0.17, 1)",
-          "type": "string"
-        },
-        "ease-out": {
-          "value": "cubic-bezier(0.22, 1, 0.36, 1)",
-          "type": "string"
-        },
-        "linear": {
-          "value": "cubic-bezier(0, 0, 1, 1)",
-          "type": "string"
-        }
-      },
-      "duration": {
-        "fastest": {
-          "value": ".2s",
-          "type": "string"
-        },
-        "fast": {
-          "value": ".4s",
-          "type": "string"
-        },
-        "default": {
-          "value": ".5s",
-          "type": "string"
-        },
-        "slow": {
-          "value": "1s",
-          "type": "string"
-        },
-        "slowest": {
-          "value": "1.5s",
-          "type": "string"
-        }
-      }
-    }
-  }
-}
