--- libs/tokens/src/tokens/space/space.json
@@ -1,74 +1,0 @@
-{
-  "sys": {
-    "space": {
-      "4XS": {
-        "value": "1",
-        "type": "float"
-      },
-      "3XS": {
-        "value": "2",
-        "type": "float"
-      },
-      "2XS": {
-        "value": "4",
-        "type": "float"
-      },
-      "XS": {
-        "value": "8",
-        "type": "float"
-      },
-      "S": {
-        "value": "12",
-        "type": "float"
-      },
-      "M": {
-        "value": "16",
-        "type": "float"
-      },
-      "L": {
-        "value": "24",
-        "type": "float"
-      },
-      "XL": {
-        "value": "32",
-        "type": "float"
-      },
-      "2XL": {
-        "value": "40",
-        "type": "float"
-      },
-      "3XL": {
-        "value": "48",
-        "type": "float"
-      },
-      "4XL": {
-        "value": "64",
-        "type": "float"
-      },
-      "5XL": {
-        "value": "80",
-        "type": "float"
-      },
-      "6XL": {
-        "value": "96",
-        "type": "float"
-      },
-      "7XL": {
-        "value": "112",
-        "type": "float"
-      },
-      "8XL": {
-        "value": "120",
-        "type": "float"
-      },
-      "MAX": {
-        "value": "999",
-        "type": "float"
-      },
-      "0": {
-        "value": "0",
-        "type": "float"
-      }
-    }
-  }
-}
