--- libs/tokens/src/tokens/state/opacity.json
@@ -1,10 +1,0 @@
-{
-  "sys": {
-    "state": {
-      "hover-state-layer-opacity": {
-        "value": "0.10",
-        "type": "float"
-      }
-    }
-  }
-}
