--- libs/tokens/src/tokens/viewport/viewport.json
@@ -1,34 +1,0 @@
-{
-  "sys": {
-    "viewport": {
-      "XS": {
-        "value": "375",
-        "type": "float"
-      },
-      "S": {
-        "value": "430",
-        "type": "float"
-      },
-      "M": {
-        "value": "768",
-        "type": "float"
-      },
-      "L": {
-        "value": "1024",
-        "type": "float"
-      },
-      "XL": {
-        "value": "1440",
-        "type": "float"
-      },
-      "2XL": {
-        "value": "2560",
-        "type": "float"
-      },
-      "3XL": {
-        "value": "3840",
-        "type": "float"
-      }
-    }
-  }
-}
