+++ package.json
@@ -244,7 +244,7 @@
     "sass": "^1.59.3",
     "sass-loader": "^13.2.1",
     "sinon": "^15.1.0",
-    "style-dictionary": "3.9.0",
+    "style-dictionary": "^4.0.0",
     "style-loader": "^3.3.0",
     "stylelint": "^16.0.0",
     "stylelint-config-prettier": "^9.0.5",
